pyproject.toml
[tool.poetry]
name = "practice-appointment"
version = "1.0.1"
description = "This is an appoinment system built using Python, Django for GraphQL API and NextJs. It is highly moduler and headless."
authors = ["Suman Sunuwar <jaycessunuwar@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
django = { version = "^4.2", extras = ["bcrypt"] }
graphene = "^3.3"
psycopg = "^3.2.1"
django-filter = "^24.3"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.2"
flake8 = "^7.1.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"

[tool.isort]
multi_line_output = 5
line_length = 79

[tool.yapf]
base_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 79
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
